feat: add AccountIdHasher for secure account ID hashing and update DebugLogMetadata with new services
