+++ src/AStar.Dev.OneDrive.Client.Core/DebugLogMetadata.cs
@@ -2,25 +2,13 @@ namespace AStar.Dev.OneDrive.Client.Core;
 
 public static partial class DebugLogMetadata
 {
-    public static class UI
-    {
-        public static class MainWindowViewModel
-        {
-            public const string Constructor = "MainWindowViewModel.Constructor";
-        }
-
-        public static class SyncTreeViewModel
-        {
-            public const string StartSync = "SyncTreeViewModel.StartSyncAsync";
-        }
-    }
-
     public static partial class Services
     {
         public static class SyncEngine
         {
             public const string StartSync = "SyncEngine.StartSyncAsync";
         }
+        
         public static class AutoSyncSchedulerService
         {
             public const string StartAsync = "AutoSyncSchedulerService.StartAsync";
@@ -29,6 +17,7 @@ public static partial class DebugLogMetadata
             public const string RemoveSchedule = "AutoSyncSchedulerService.RemoveSchedule";
             public const string AutoSyncTriggered = "AutoSyncSchedulerService.AutoSyncTriggered";
         }
+
         public static class DeltaPageProcessor
         {
             public const string ProcessAllDeltaPagesAsync = "DeltaPageProcessor.ProcessAllDeltaPagesAsync";
